Add support for invoking documentation tool
This CL adds a new flag, --generate-documentation.
The next argument on the command line should point to a
documentation tool (like Dokka or javadoc), and the remainder
of the command line are all passed as arguments to that
documentation tool. (You can also specify --no-docs earlier
on the command line to disable this; this is useful for a
general metalava integration where individual tasks can add
in local arguments, such as --no-docs to turn off documentation
without needing to invoke a different target.)

In unrelated news, this CL also fixes
ndroid-ktx/issues/366:
Metalava tracks wrong parameter defaults with @JvmOverloads

Test: Unit tests included
Change-Id: I4f04054d5a31924167ff6213294ed5edb26e8b1a