Add annotation extraction for typedef constants
Metalava contained the code for extracting annotations, but
this was the older code using its own analysis (calling into
the extractions code in AGP); not connected to the current
codebase.

With moving annotations into the android.jar file, we don't
want these annotations redundantly declared in the external
annotations database. Furthermore, we want to do filtering
based on the ApiPredicate in the loaded codebase, not based
on the older API filter mechanism.

This CL rewrites the extraction code to run locally on the
metalava visible class items instead. Also cleans up
a few other things and properly handles Kotlin.

Test: Unit tests included
Change-Id: I1c65c6ea68cb1848a512e506475427c0d1ba1f31