Remove most libcore annotations from metalava/manual.
Most of the annotations can be removed without making any difference
to the contents of metalava-api-stubs-docs.jar because they are
already present in //libcore/annotations/ojluni.jaif.

This change does make four changes to the contents of that jar, which
are described below.

A small number of libcore annotations are left here. These are all
present in ojluni.jaif, but metalava is not picking them up correctly,
because of an issue around its handling of annotations on return types
(b/111116803). They are left here as a temporary measure to work
around that issue.

Character.toChars now appears in the stubs as
  public static char @androidx.annotation.RecentlyNonNull [] toChars(int codePoint)
instead of
  @androidx.annotation.RecentlyNonNull public static char @androidx.annotation.RecentlyNonNull [] toChars(int codePoint)
The new version is the correct syntax for a non-null array of primitive char.

Map.put now appears in the stubs as
  @androidx.annotation.RecentlyNullable public V put(K key, V value);
instead of
  @androidx.annotation.RecentlyNullable public V put(@androidx.annotation.RecentlyNonNull K key, @androidx.annotation.RecentlyNonNull V value);
The new version is correct since many maps do support null
values. (Well, actually the most correct thing is the equivalent of
libcore's @NullFromTypeParam, when that exists.) Similiarly for
putIfAbsent and replace.

Test: make metalava-api-stubs-docs
Change-Id: Ibc67b79decbaa8c6a4893d6f87a0a7c87cb88c4d