111216907: Private stub annotation classes breaks compilation
This CL updates metalava such that it *only* packages
the @RecentlyNullable and @RecentlyNonNull annotations
as package private stub annotations; the rest of the
annotations are packaged as external annotations (which
was the case prior to P anyway).

It also turns off type-use annotations, since they are
not properly supported anyway; their restoral is tracked
in issue 111253910.

Fixes: 111216907
Test: Unit tests updated

Change-Id: Ia9b85bf75e143ae3dd622300f1731a30df2a16cd
Merged-In: I1d454024aeaee60daa8c717d40e352300ff768cc