Misc metalava improvements
(1) Fix pretty printing default values such as lambas

This CL cleans up the code formatting a bit (used to emit
things like annotation defaults, default values in signature
files, etc), and in particular fixes it such that it does not
include newlines in signature files, omits implicit
annotations, etc.

(2) It fixes compatibility checking with signature files.
The old signature files are incomplete (for example, omitting
some inherited members from hidden super classes), which meant
the compatibility check was detecting the presence of these
in the codebase as newly added methods (and there are similar
issues around modifiers).

(3) It makes the --check-compatibility and --migrate-nulls
flags take the respective signature files as arguments, and
deprecates the previous --previous-api and --current-api flags
since it wasn't really obvious how they applied.

(4) It adds a check which flags unexpected constants returned
from methods annotated with a typedef.

To do this it looks at methods annotated with a typedef annotation,
and checks all the return statements in the method; if it finds
that it returns a constant not recorded in the typedef it
flags this as a warning; the usual cause is updating the code
but forgetting to update the typedef.

Some misc warnings cleanup and other code cleanup.

Test: Unit tests included
Change-Id: Ia31f60adc45dfc3b50e3e2d6259c5a916384b1d7

Fix API comparisons

Also support args to --check-compatibility

Change-Id: Ic019c1e5ea02b86faa33aee376b61bcbb5648a53