Ensure that stub files contain inherited fields
In system-api and test-api signature files, only
annotated elements are included. This is not the case
for stub files, where we always include all the elements,
but some special handling around inherited fields caused
constants to be filtered out from signature files; this
CL fixes this and adds a regression test.

Test: Unit test included
Change-Id: Ic83398d2736fc8b0551a524a0533a9f900c5903c