Remove unused code
This CL removes a lot of leftover unused code from an earlier
mechanism in metalava were it created a whole separate codebase
instance by filtering a source codebase. The idea was that this
would be faster (for subsequent operations on the codebase such
as iterating it for each writer (signature, stubs, proguard, etc)
as well as the various other codebase iteration (api enhancer and
so on) but in practice this didn't work well.

Also fix various warnings.

Test: Covered by existing tests
Change-Id: Ic5582c233ad30897996178ced7ecba792dcbe370