Add support for multiple levels of compatibility checks
When verifying the API against signature files, we need to
apply different sets of checks when comparing the codebase
against current.txt versus checking it against a previously
released API.

In doclava, this was handled by running doclava multiple
times and passing in various combinations of -error <id>
and -hide <id> flags to configure things. In metalava, this
is instead pushed into the tool; it now supports sets of
errors and can switch between these based on the type of
compatibility check to be run.

The --check-compatibility flag has been changed to specify
both the API to be checked, and the release mode. This
means that there are these supported variations right now:
  --check-compatibility:api:current
  --check-compatibility:api:releasedd
  --check-compatibility:removed:current
  --check-compatibility:removed:released

The first two are checking the public API, against either
the current.txt file or the last released public API (e.g.
28 at this point). The second two are checking the removed
API.

As before, this flag should be followed by a path pointing
to a signature file.

This CL also includes a fix to make system API checking
work against the currently checked in system-current.txt.
That file was generated by doclava, and is missing the class
android.telephony.ims.feature.ImsFeature.Capabilities.
That class is included in the signature file by metalava
but not by doclava, so it looks like it's an incompatible
change (newly added class), but this is a class which should
have been included in the signature file by doclava (it's
in the API and referenced for example from the
MmTelFeature.MmTelCapabilities constructor in doclava's
signature file as well) and therefore it's deliberately
excluded from the API check.

Also includes a workaround for misconfigured sourcepath
options where pointing to a directory containing
source roots rather than being source roots themselves.

Test: Unit tests included
Change-Id: Ib742904c7ad6273a92c4f99a855ec5ae73512e0a