Allow suppressing issues with specific error messages
There have been some cases where the API changed incompatibly,
for example changing the initial value of a field.

When this happens, you can suppress the warning with
@SuppressLint("ChangedValue") -- but that would effectively stop
flagging changed values for this constant forever, and you probably
want to make sure it doesn't change to some other value in the
future.

This CL adds a refinement to the @SuppressLint mechanism where
you can specify the exact error message which must also be
matched:

  @SuppressLint("ChangedValue: Field myfield has changed value from 1 to 42")
  public static final int myfield = 42;

In the above, the error will be suppressed only if the new
compatibility check failure message is identical to the message
after the ":" (optionally followed by a space).

This CL also cleans up the compatibility checking code a bit, and
updates the failure message when there are compatibility problems
to list which compatibility check and signature file the check
failed against (since a single metalava invocation can typically
be checking 4 different signature files with different
release modes (and therefore configured rules), so it's important
to make it clear which API failed the test.

Test: Unit test included
Change-Id: I5631643af8a84c774cc0bc003c4ccefeb54894b3