Add support for --dex-api-mapping
This CL adds support for --dex-api-mapping to metalava, which
corresponds to the -apiMapping flag in doclava (which it also
respects). This emits a file containing the dex signatures for
all methods and fields (not classes) in the API, and on the
following line, the file and line number where the modifier
list of the field or method begins.

Some other cleanup.

Test: Unit tests included
Change-Id: I2ce302d46f921b10fb408c0da5e6e65e291ffc50