Add support for --update-api
This CL adds an option, "--update-api", which tells metalava that
it is invoked as part of an --update-api task, and therefore to
ignore various other action flags, such as --check-compatibility.
When --update-api, only the source description flags and the
signature generation flags are respected.

This makes it easier to implement the update-api task in the
build system; we don't have to clone the task with a subset of
the flags (which may themselves change over time). And it is
similar to the --no-docs flag which has a similar effect of
canceling all the other documentation flags.

Test: Unit test included
Change-Id: Ie60707a4639dd8f29047deb380f4f55f86afccdc