Warn about @SystemApi/@TestApi not marked with @hide
This can lead to accidentally exposing API meant to
be internal. The check is currently turned off (severity
is hidden) since there are ~160 violations in the source
tree; it can be turned on with --error UnhiddenSystemApi.

Also unifies several separate full tree traversals for
API checks into a single one as mentioned in one of the
TODO's, and switches the issue type for the system service
check from RemovedField to RequiresPermission.

Test: Unit tests updated
Change-Id: I12119002686eac9451993d19b393a4fc453bc683