Work around metalava recursion issue
As part of introducing a new internal "intra-core" API surface
we need an annotation (libcore.mmodule.IntraCoreMModuleApi).

Because we're in the core libraries (i.e. the bottom of the Java
dependency tree) there's nothing lower we can depend on so the
annotation must annotate itself to be included in its own API
surface.

This change fixes a recursion error in metalava that results when
using libcore.mmodule.IntraCoreMModuleApi.

Bug: 113148576
Test: build
Change-Id: Iac7220a6b80844b662ed1858bb0f9fe8ec63589b