Fix a few issues in compatibility checks against v2 signature format
This CL fixes a few bugs in the compatibility check code when
handling v2 signature files. For example, in v2 we now explicitly
include nullability, including for implicit nullness (e.g. we
know that an annotation type member cannot be null; therefore
the signature file marks it as non null, but the compatibility
check did not consider this so the lack of a @NonNull annotation
there triggered a "you cannot remove @NonNull" warning.)

Test: Unit test included
(Also manually verified ./gradlew checkApi in frameworks/support/
with metalava enabled.)
Change-Id: I7dabdcc41868ab1594e0e7a3f63289ba4fb2ad9d