Add support for --show-single-annotation
This CL adds support for a new flag, --show-single-annotation, similar
to the existing --show--annotation flag, but with the distinction that
it only unhides the exact (single) element that it is annotating; it
does not apply to children.

This is useful if you want to mostly hide a whole package and then be
able to only include members (methods, fields, inner classes) that you
are *explicitly* annotating with this show annotation.

Note that default constructors are implicitly included when the
surrounding class is annotated with one of these new annotations. If
you don't want that, explicitly define the constructor and hide it.o

Test: Unit test included
Change-Id: If07985f6625ae1b4e6e40f61932f86ef4cdc715e