Add metalava support for merging show/hide annotations.
Previously, metalava only merged annotations that would appear on the
API from external codebases. With this, it also merges show/hide
annotations. This is required by libcore's mainline modules to allow
us to keep mainline module show/hide annotations in external stub
files.

This is handled via a separate flag, because the merging has to happen
earlier in the process, because the show/hide affects the behaviour
of the `analyzer.computeApi()` step as well as the merging of API
annotations themselves.

Bug: 115746226
Test: `./gradlew test`
Test: `make core-platform-api-stubs` with local modifications to trigger this
Test: `make api-stubs-docs`
Change-Id: I0b9392b204ea61e6c33e450239b5cedeb67ce986