Rewrite documentatation stubs to be fully qualified
When running doclava to generate javadocs from the
documentation stubs, it generates surprising unresolved
link errors, even for symbols that are clearly imported
in the same compilation unit.

This CL changes metalava such that it rewrites all
javadocs in the documentation stubs to be fully
qualified instead, to work around this issue.

For now, metalava does *not* flag unresolved symbols;
it was generating thousands of these, and in various
spot checks it looks like the documentation is actually
wrong, so we'll need to clean that up first.

Test: Unit tests included
Change-Id: I6dce30595b38aee8fc285e45aa4925fe6dbd186d