Add nullability annotation validation to metalava.
The main functionality here is a generalized validator which can check
any list of classes. There's a flag which causes all classes in merged
java stub files to get validated, which is what libcore needs (but it
is designed to make it easy to add alternative mechanisms for
validating other classes).

This considers annotations that don't make any sense to be errors, and
missing annotations on the validated classes to be warnings. By
default, it fails if there are errors and writes out warnings.

This doesn't yet look at annotations on arrays or on type parameters,
as metalava support for TYPE_USE is needed first (see the comments in
the source).

Test: ./gradlew test
Test: making a new libcore target using this
Bug: 73448108
Change-Id: Iad53befbe28a16ab76d260f61f84fbada1eb74df