117553509: Metalava prints empty lines on successful runs
Part of the challenge is that the JDK parsing infrastructure
registers a default XML error handler which prints to
stdout by default so this CL adds some central XML
parsing utilities which registers an error handler which
wraps the print streams.

This CL also adds testing infrastructure which redirects
the output and error streams and throws an error if any
code prints directly to stdout instead of going through
the dedicated print writer.

Change-Id: Iaf9d26d6388845dc1015487be55309d0aee64e9d
Fixes: 117553509
Test: Now enforced by unit test infrastructure