Treat accessing hidden APIs as an error
The current behavior of doclava and metalava is
to force hidden classes as public if they are accessed
from public API (for example as a parameter type).

This has resulted in accidentally exposing API
in the past.

With this CL, this instead treated as an error.
Now developers should explicitly unhide API to
make it visible.

(There were various existing warnings about this,
but they are warning severity or in some cases
turned off, and this new error (a) also applies to
package-private APIs, and (b) is a single issue type
to check all hidden access instead of having different
issue types to configure for each types of API access.)

Also, add --format=vN flag to set output format.
There were various flags controlling aspects of the
output, but this flag encapsulates the various
formats:
1) doclava
2) new format, but with @Nullable/@NonNull annotations
3) new format, but using Kotlin syntax for nullness

Test: Test included
Change-Id: I4fef0588a2fd55fe6a5f67f6baffe8071fe565b6