Don't show @RecentlyNullable in signature files
When generating the stub files, we deliberately change
@Nullable/@NonNull to @RecentlyNullable/@RecentlyNonNull if the API
was not also marked as @Nullable/@NonNull in the previous API level.

However, when switching to signature file format 2, this makes
@RecentlyNullable show up in the signature files.  We don't want that,
so defer the nullness migration step until after the signature files
have been written, but before the stub files are written.

This wasn't apparent until now because in format 1, annotations don't
show up in signature files, and until recently, format 2 also used
Kotlin-style nulls.

Test: Unit tests updated
Change-Id: Icd2311504b59e29535255a65118ec14968248fb0