119637705: Fix API check to catch added system methods
For system- and test-api signature files, the signature
files contain only the *delta* between the system/test api
and the real Android API. That means that during API check,
it has to work a bit harder: if a method is moved from
the system API and exposed in the public API, it will now
disappear from the system API, but it's not an incompatible
change, so metalava searches for inherited methods in the
base public/API as well.

There was a bug in this code, which this CL fixes: the
inherited method lookup would sometimes return self
rather than inherited methods.

Change-Id: Ice3e2e76b9b58faef182c46718c12c48ae721ef4
Fixes: 119637705
Test: Unit test included