Support non-migration annotations
Until Android P, the SDK did not contain nullness annotations,
so in Android P all annotations were marked as @RecentlyNullable
or @RecentlyNonNull.

This CL fixes things such that as of master/Q, when a nullness
annotation on an API is not new, the annotation is instead mapped
to a non-@Migrate annotation as far as the Kotlin compiler is
concerned: android.annotation.{NonNull, Nullable}.

Also makes some internal annotation processing more efficient.

Test: Unit tests included
Change-Id: Ia2144b5142cb0ce5e5303a5c72fb9c7c6e3a69d7