119859108: Require default constructors to explicitly opt in
With the --show-single-annotation flag, APIs can be annotated
to be hidden without inheritance. This was originally implemented
to explicitly exempt default constructors.

Issue 119859108 requested this behavior to change; now default
constructors must be explicitly added to the source code and
annotated with the API annotation instead to be part of the API.

Test: Unit test updated
Change-Id: Id1d54733ca7fbea20b47d7d3bcfa7d7e85e6e93b
Fixes: 119859108