120005399: Remove retention policy and element type
When not generating class retention annotations into stubs
(such as nullness annotations, via --include-annotations)
the code to compute the retention of annotation classes
was not working correctly in scenarios where the
java.lang.annotations are not part of the source tree (e.g.
when not building the SDK).

Also fix a bug in annotation sorting in stub files.

Change-Id: I3462e514844f7ee9d06a3a6080d43fbbf0df07f6
Fixes: 120005399
Test: Unit test included