Metalava baseline support
This CL adds support for "baselines" to metalava, similar to
Android Lint's baselines. This lets you record the current set
of warnings, which are then checked in as text files, and from
that point forward, only warnings and errors not found in the
baseline are reported. Therefore, the current set of problems
in the codebase form a "baseline" against which newer issues
are compared.

Test: Unit tests included

Change-Id: I8ebee2aaef269925f2a1b0b1e5bfb17f05ee324d