Fix bugs related to baselines and clean up error output
First, this CL tweaks the baseline format such that
the API identifier and the error message are on separate
lines, making it easier to read the text file. It also
fixes a problem with --update-baseline such that it now
works correctly even if the baseline already exists, and
adds dumping statistics in verbose mode.

Second, it updates the error and reporter classes such
that we no longer include the numeric values for errors
next to id's (since these were sometimes confused with
line numbers), and add in categories.

Test: Existing unit tests updated
Change-Id: I34212c006c66e0e605973e598719cd5482f9ea56