Fix --stub-packages handling
This CL fixes handling the --stub-packages argument;
correctly handle getting called with "*" as a package
suffix. Also limit signature file and error reporting
to the stub packages, and properly document this
argument.

Also update the default baseline location such that
it defaults to the api/ folder, if present, since this
is a convention in the platform for the related
signature files.

Test: Unit tests added
Change-Id: I2394c60a413aadec2f1870184686a62a0ddcb664