Include package private classes referenced from public API
If the code violates the ReferencesHidden check (which some
libraries do, and suppresses the check to make the build pass),
the stubs won't compile: they reference types that metalava
does not generate stubs for.

This CL changes that: it will now generate these stub files
too (but only for the class itself; none of the members).

It also fixes a bug around adding casts for super constructor
calls: if the super type is a type parameter, don't use
the erased type, use the type argument to the super
constructor.

Test: New unit tests included
Change-Id: I8bc09cf0dc321f41653830afa2db130de1b8dbd5