122916999: Compat check with *current: ensure no file diffs
When running compatibility checks against current.txt,
system-current.txt etc, also require that the signature
file itself has not changed. This is necessary since there
are many changes you can make which are visible in the
signature file but compatible, and we want these to show
up immediately such that they don't accumulate and show up
as diffs for the next engineer who makes an incompatible
change.

Fixes: 122916999
Test: Unit test included
Change-Id: Iae62cd1ccb706e888b83d9fe41395d4775795e57