123140708: Always include spaces in type argument lists in jdiff output
This CL fixes CTS tests that depend on string equality between
the JDiff type descriptions and their own internal pretty-printer
based on reflection lookup of the API; the pretty printer always
emits a space after a comma in type lists whereas metalava's
signature file to JDiff output converter would preserve whatever
was in the signature file, which as of v2 omitted spaces.

This CL forces the output to always have spaces when converting
to JDiff output.

Test: Unit test included (plus of course the failing CTS test)
Change-Id: Icbcb3067423cb5fa259e18e028704c27c48953e4
Fixes: 123140708