Fix type comparison to ignore whitespace
When comparing type signatures,	spaces should not be significant,
eg. Map<X,Y> ==	Map<X, Y>. This	was already mostly the case, except
for when comparing types from one signature file with types from
another, as is the case when doing compatibility checks with show
annotations (such as the system API).

Test: Unit tests included
Change-Id: Icec1a679cfb9245905581bc0fe9074745f835038