Generate @apiSince and @deprecatedIn instead of literal text
This is intended for doclava and other tools to do their
own rendering of the doc content to describe a deprecation
level.

Also renames the @since tag to @apiSince to make it less
ambiguous and to not conflict with existing @since tags
that are present in various docs.

Also starts emitting @apiSince into package.html docs,
such that there is a concept of a package API level; this
is the lowest API level for any class in that package.

Also makes sure we don't emit @apiSince for SystemApi
docs, since we don't have accurate historical information
for SystemApi and TestApi sources.

Also starts writing @apiSince tags even when the since tag
is 1 (e.g. for the APIs added from the beginning). This
was omitted for optimization purposes earlier but is added
back to make the doc generation task easier.
I also added some optimizations for these two new tags since
they can be merged more quickly with some special handling
since they never appear in existing sources and can always
be listed last in the docs.

Finally, various fixes to the API lookup data base
generation; this makes the generated database more
closely mirror what's in android.jar, and, importantly
for doc generation, properly tracks in-development APIs,
such that generated docs at the moment shows up as "Q"
instead of "28".

Test: Unit test updated
Change-Id: If25a8075dc1bb2ace184d1b4d6f19717fae2bc83