Refactoring: Rewrite links in Metalava instead of carrying patches.
Some .java files in libcore/ojluni contain links that go outside of
{@docRoot}. Before this CL topic, libcore was carrying 208 patches to
these links. This CL topic adds logic to rewrite the links at doc
generation time instead; this allows us to drop the libcore patches in
the libcore CL that is part of this CL topic.

This CL introduces a new metalava feature parameter
--replace-documentation with the following arguments:
   1. a list of packageNames, separated by :
   2. a regex
   3. a replacement (may contain backreferences $1, $2, ...)

Each such parameter results in all matches of the given regex to be
replaced with the given replacement for all Items from the listed
packages and their sub-packages.

This CL topic uses the new feature to express the task of rewriting
specific links in ojluni documentation as a --replace-documentation
parameter (see frameworks/base CL in this CL topic) which has the
advantage that metalava doesn't need hard-coded knowledge about
ojluni specifics.

Ideally, the regex would be parameterized by the nesting depth of
the package currently being processed because we expect exactly
depth+1 occurrences of "../" at the start of the relative URL to
ascend to one level above {@docRoot}. To simplify the new
functionality, the regex is not parameterized and could therefore
also match additional links; but in practice, the requirement
for "../platform" or "../technotes" to occur in the link is
sufficiently specific that the more lenient regex does not match
any false positives; I've verified this by comparing the output
against a version that performed such parameterization.

Bug: 129765390
Test: Checked that this CL topic does not affect the docs generated by
      make ds-docs offline-sdk-docs
      (in other words, the docs generated before vs. after this CL
      topic are equivalent).
Test: croot tools/metalava ; ./gradlew test

Change-Id: If64ebe6bddd8c24be4137736c37c4f0afce5aec8