Fix Metalava not fixing summary lines that contain "e.g. ".
javadoc cuts off the summary line after the first ". ", which can
lead to unintended results when the first ". " is part of "e.g. "
or similar. This occurs in 225 .html places in the Android
documentation, for example the summary of the constant
TEXT_ALIGNMENT_CENTER in [1] reads:

   Center the paragraph, e.g.

when it should be reading:

   Center the paragraph, e.g. ALIGN_CENTER.

Metalava had logic to work around this by replacing

  "e.g. " -> "e.g.&nbsp;"

however this logic is part of the typo fixing logic in Metalava, and
that logic is currently disabled [2].

It'd probably be reasonable to re-enable the typo-fixing, but the
fact that it's currently disabled, and that two other places have
copied the command line including this argument, suggest that it's
not safe to rely on it remaining enabled. Since the code delivers
significant value with (as far as I can see) no downside / no false
positives, this CL changes the code such that the fixing of "e.g. "
can no longer be disabled via a command line switch.

Bug: 131580537
Test: Ran "make ds-docs offline-sdk-docs" before and after this CL.
      Out of the 225 files affected, I inspected approximately 25
      and confirmed that each of them had previously suffered from
      the problem described above and that it was now fixed.
Test: ./gradlew test

[1] https://developer.android.com/reference/android/view/View
[2] https://android.googlesource.com/platform/frameworks/base/+/refs/heads/master/Android.bp#1157

Change-Id: Id0d982c3a7dadc38250e457381be926173cd1bb2