Add ClassItem.isDefined() to query whether class was defined in API
When parsing metalava creates an instance of ClassItem for every class
referenced in an API specification, irrespective of whether the class
is defined in it or not. The isDefined() method allows a client to
determine if the class represented by ClassItem was defined.

This is needed by cts/tests/signature test packages that parse an API
specification to extract the names of classes that should not be
available. If undefined classes are included in that list then it
breaks the tests.

Test: atest -p cts/tests/signature
Change-Id: Ibd8d4f6f902978069062457e3972968cad3ab9f0