Simplify constructor selection in stubs
Constructor selection was previously done on the unfiltered constructor
list. This made the code more complex and could result in constructors
that reference excluded types being present in the generated stubs. By
selecting from the already filtered constructors the code can be
greatly simplified and the resulting stubs will not reference excludeed
types.

It does change the behavior of some of the tests:
* CoreApiTest
  * 'Hidden with package javadoc and hiding default constructor explicitly'
    Removes the /**@hide*/ comment that was copied over from the
	original source. That is because the constructor is no longer copied
	over and instead was created.
* StubsTest
  * A number of tests were affected as instead of reusing a hidden
    constructor with parameters and throws from the source it creates a
	new no args constructor.

This should have no adverse effects because in the case when a
constructor that referenced excluded types was output it could never
actually have been used because it was marked as package private. So,
it is safe to replace it with a default package private constructor
as it did previously when it could not find a suitable constructor.

Bug: 141219826
Test: ./gradlew test + m checkbuild + manual inspection of affected classes
Change-Id: I62be87d47124d135a50a4c0b98f55c9d63dc1db5