Improve handling of visibility level
Merged separate bits for each possible level (which can lead to
multiple levels being set) into a group of bits that can only specify a
single level.

Removed some unused methods: set(String), bit(String)

Did not remove the asAccessibleAs(ModifierList) method as while that is
unused it is referenced from some TODOs so I left it in.

Removed the expression "(flags or PRIVATE) != 0" as that always
evaluates to true because PRIVATE is not 0.

Test: ./gradlew test && m droid
Bug: 141229417
Change-Id: I193850bd2972d07d1fc6accefdf18ee0e98751d0