Transform API model before merging annotations/validating/checking/etc.
After loading and computing the API an additional transform is applied
to the API module. Any class members that are part of the API but
declared on classes that are not part of the API are copied onto the
hidden class' subclasses. That transformation must be done before
validation of the API as otherwise the validation is being done on an
incomplete API. It also needs to be done before merging in qualifier
annotations as otherwise it is impossible to add annotations to those
methods because the annotation merger ignores hidden classes.

This change detected some missing nullability annotations on
CharSequence methods implemented by the AbstractStringBuilder hidden
class but exported by StringBuilder. Previously, the errors were not
detected because the affected methods from AbstractStringBuilder were
not present in the API model.

Bug: 143700871
Test: m checkbuild && ./gradlew test
Change-Id: Idbebb0bf380a65e37f232d2ab5ec240a77acddd7