Ensure annotations are merged/migrated on System/TestApi stubs
Removed filter that prevented annotations being merged into and
migrated on classes that were generated as stubs but which were not in
the API signature, e.g. classes that are part of the current API but
are not in the system-current.txt or test-current.txt.

That should not cause a problem because the filters will be applied
before the stubs and APIs are output.

Verified that RecentlyN... annotations were not present on the
ConcurrentHashMap stubs in test-api-stubs-docs and
system-api-stubs-docs before my change but were afterwards.

Also, verified that the new tests that I added failed without my
changes and worked with them.

Bug: 144114248
Test: m checkbuild && ./gradlew
Change-Id: If28bf262ec5f94f8e421afbeb837e54fdb2b6ec0