Annotations with default attributes are handled
This change fixes a bug that APIs with an annotation having default
attributes are not filtered when the annotation is listed in
--show-annotation along with attribute values. For example,

public @interface Api {
    enum Type { A, B}
    Type type() default Type.A;
}

public class SomeApiClass {
    @Api
    public void foo() {...}
}

The method SomeApiClass#foo was not filtered with
--show-annotation Api(type=Api.Type.A)

This was because when converting PsiAnnotation (from the intellij lib)
into PsiAnnotationItem in metalava, only the declared attributes were
considered. When there is no attribute declared (with the expectation
that default values will be used instead), metalava does not actually
try to get the default values and thus emits an annotation without
attribute, which fails to match with the filter spec.

This change fixes the problem by reflectively enumerating attributes
from the annotation class and using findAttributeValue() to retrieve the
default value when the attribute is not declared.

Note: The stub-only NonNull and Nullable annotations are added with
from() and to() attributes, because libcore.annotation.NonNull|Nullable
which is their source of truth are with the attributes.

Bug: 145647640
Test: ./gradlew test the ShowAnnotationTest is passing
Change-Id: I265a5905743766d2a51dd55b7dc50bfe6f57ceb3