Avoid ConcurrentModificationException
When iterating over classes in a package, topLevelClasses() is called.
Previously, that method was directly using the 'classes' field which is
a mutable list. However, in certain cases (explained below), the list is
mutated while traversing over it, causing
ConcurrentModificationException. This change fixes the problem by
copying the list (but not the contained classes) in topLevelClasses().

The concurrent modification can happen when following conditions are all
met:

1) A Java file is added to the source files list
2) A java class is from the classpath
3) 1) and 2) are in the same package
4) 1) is tagged with an annotation whose attribute values are from 2)

Then while parsing 1), addClass() is called for 2) which modifies the
classes field in the containing package.

Bug: 146398156
Test: m with aosp/1186065
Test: ./gradlew test passes, and fails without the change in
PsiPackageItem.kt

Change-Id: Ie0597b07ce2f20ae51178d15b4d054b5f10749f9