Making `hidden` be consistent regardless of file iteration order
PsiBasedCodebase's `hidden` check was dependent on iteration order:
if the package-info.java was not parsed until after the class's java file,
then the package would be not found and would not be considered to be hidden.

However, ApiAnalyzer.propagateHiddenRemovedAndDocOnly already sets `hidden`
on classes having packages marked as "@hide", but it also does a few other
checks involving hasShowAnnotation() too.

So, we can remove the check in PsiBasedCodebase and rely on the check in ApiAnalyzer

Bug: 146077896
Test: ./gradlew test --tests "com.android.tools.metalava.ShowAnnotationTest*order*"

Change-Id: Ia13f43a9dd7c5dd0e8e4094908fc1c58aac5eb94