Do not exempt sourcepath from sandboxing.
This CL prevents the exemption of sourcepath from the sandbox if
enabled. It also bypasses violations when isDirectory or isFile are
used when traversing the sourcepath, and only reports violations
when files are read under sourcepath.

Test: built api-stubs-docs and framework-doc-stubs and verified there
are violations under the sourcepath.

Bug: b/156613606
Change-Id: I66fdf98d9a40b13754b2f613111bc6ff3794d83e
Merged-In: I66fdf98d9a40b13754b2f613111bc6ff3794d83e